feat:  Add arguments 'offset' and 'n_period' to function header mp_compute_indicator
Allow for better control of the the number of LPPLS fits performed. Rather than performing all end dates corresponding to the indices length(x)-1, ..., window_size-1 of the array self.observations, we allow for explicit control of the maximum end date and the number of end dates for which the fits are performed. The maximum end date is determined by 'offset', and the number of date for which the fits are performed is defined by 'n_periods'.  I.e. if a price array self.observations of length x is given, the fits required for the calculation of the confidence indicator will be performed for all end dates corresponding to the array's indices length(x) -1 - offset, length(x) -1 - offset - 1, ..., length(x) - 1 - offset - n_periods